Added Fmnist and training code
Implemented some training code for the FNN and used the mnist dataset (easy to change to another). I can not push the Fmnist to the repo due to size..  Next steps are improve the testing with softmax+cross entropy? But the simple tests came out pretty decent (75-80% accuracy) But need to test bigger samples.